example step 1: extract and rename v53 service
